+++ src/search/pdbs/pattern_collection_generator_combo.h
@@ -3,17 +3,20 @@
 
 #include "pattern_generator.h"
 
+#include "../options/options.h"
+
 namespace pdbs {
 /* Take one large pattern and then single-variable patterns for
    all goal variables that are not in the large pattern. */
 class PatternCollectionGeneratorCombo : public PatternCollectionGenerator {
-    int max_states;
+    options::Options opts;
+
+    virtual std::string name() const override;
+    virtual PatternCollectionInformation compute_patterns(
+        const std::shared_ptr<AbstractTask> &task) override;
 public:
     explicit PatternCollectionGeneratorCombo(const options::Options &opts);
     virtual ~PatternCollectionGeneratorCombo() = default;
-
-    virtual PatternCollectionInformation generate(
-        const std::shared_ptr<AbstractTask> &task) override;
 };
 }