+++ src/search/pdbs/pattern_collection_generator_hillclimbing.cc
@@ -26,6 +26,7 @@
 #include <cassert>
 #include <iostream>
 #include <limits>
+#include <string>
 
 using namespace std;
 
@@ -113,7 +114,8 @@ static vector<vector<int>> compute_relevant_neighbours(const TaskProxy &task_pro
 
 
 PatternCollectionGeneratorHillclimbing::PatternCollectionGeneratorHillclimbing(const Options &opts)
-    : pdb_max_size(opts.get<int>("pdb_max_size")),
+    : PatternCollectionGenerator(opts),
+      pdb_max_size(opts.get<int>("pdb_max_size")),
       collection_max_size(opts.get<int>("collection_max_size")),
       num_samples(opts.get<int>("num_samples")),
       min_improvement(opts.get<int>("min_improvement")),
@@ -252,9 +254,9 @@ pair<int, int> PatternCollectionGeneratorHillclimbing::find_best_improving_pdb(
             improvement = count;
             best_pdb_index = i;
         }
-        if (count > 0) {
-            utils::g_log << "pattern: " << candidate_pdbs[i]->get_pattern()
-                         << " - improvement: " << count << endl;
+        if (count > 0 && log.is_at_least_verbose()) {
+            log << "pattern: " << candidate_pdbs[i]->get_pattern()
+                << " - improvement: " << count << endl;
         }
     }
 
@@ -304,8 +306,10 @@ void PatternCollectionGeneratorHillclimbing::hill_climbing(
     const TaskProxy &task_proxy) {
     hill_climbing_timer = new utils::CountdownTimer(max_time);
 
-    utils::g_log << "Average operator cost: "
-                 << task_properties::get_average_operator_cost(task_proxy) << endl;
+    if (log.is_at_least_normal()) {
+        log << "Average operator cost: "
+            << task_properties::get_average_operator_cost(task_proxy) << endl;
+    }
 
     const vector<vector<int>> relevant_neighbours =
         compute_relevant_neighbours(task_proxy);
@@ -329,7 +333,9 @@ void PatternCollectionGeneratorHillclimbing::hill_climbing(
       guaranteed to be "normalized" in the sense that there are no duplicates
       and patterns are sorted.
     */
-    utils::g_log << "Done calculating initial candidate PDBs" << endl;
+    if (log.is_at_least_normal()) {
+        log << "Done calculating initial candidate PDBs" << endl;
+    }
 
     int num_iterations = 0;
     State initial_state = task_proxy.get_initial_state();
@@ -342,14 +348,19 @@ void PatternCollectionGeneratorHillclimbing::hill_climbing(
         while (true) {
             ++num_iterations;
             int init_h = current_pdbs->get_value(initial_state);
-            utils::g_log << "current collection size is "
-                         << current_pdbs->get_size() << endl;
-            utils::g_log << "current initial h value: ";
-            if (current_pdbs->is_dead_end(initial_state)) {
-                utils::g_log << "infinite => stopping hill climbing" << endl;
+            bool dead_end = init_h == numeric_limits<int>::max();
+            if (log.is_at_least_verbose()) {
+                log << "current collection size is "
+                    << current_pdbs->get_size() << endl;
+                log << "current initial h value: "
+                    << (dead_end ? "infinite" : to_string(init_h))
+                    << endl;
+            }
+            if (dead_end) {
+                if (log.is_at_least_normal()) {
+                    log << "Initial state is a dead end. Stop hill climbing." << endl;
+                }
                 break;
-            } else {
-                utils::g_log << init_h << endl;
             }
 
             samples.clear();
@@ -365,8 +376,10 @@ void PatternCollectionGeneratorHillclimbing::hill_climbing(
             int best_pdb_index = improvement_and_index.second;
 
             if (improvement < min_improvement) {
-                utils::g_log << "Improvement below threshold. Stop hill climbing."
-                             << endl;
+                if (log.is_at_least_normal()) {
+                    log << "Improvement below threshold. Stop hill climbing."
+                        << endl;
+                }
                 break;
             }
 
@@ -375,9 +388,13 @@ void PatternCollectionGeneratorHillclimbing::hill_climbing(
             const shared_ptr<PatternDatabase> &best_pdb =
                 candidate_pdbs[best_pdb_index];
             const Pattern &best_pattern = best_pdb->get_pattern();
-            utils::g_log << "found a better pattern with improvement " << improvement
-                         << endl;
-            utils::g_log << "pattern: " << best_pattern << endl;
+            if (log.is_at_least_normal()) {
+                log << "found a better pattern with improvement " << improvement
+                    << endl;
+                if (log.is_at_least_verbose()) {
+                    log << "pattern: " << best_pattern << endl;
+                }
+            }
             current_pdbs->add_pdb(best_pdb);
 
             // Generate candidate patterns and PDBs for next iteration.
@@ -389,30 +406,39 @@ void PatternCollectionGeneratorHillclimbing::hill_climbing(
             // Remove the added PDB from candidate_pdbs.
             candidate_pdbs[best_pdb_index] = nullptr;
 
-            utils::g_log << "Hill climbing time so far: "
-                         << hill_climbing_timer->get_elapsed_time()
-                         << endl;
+            if (log.is_at_least_verbose()) {
+                log << "Hill climbing time so far: "
+                    << hill_climbing_timer->get_elapsed_time()
+                    << endl;
+            }
         }
     } catch (HillClimbingTimeout &) {
-        utils::g_log << "Time limit reached. Abort hill climbing." << endl;
+        if (log.is_at_least_normal()) {
+            log << "Time limit reached. Abort hill climbing." << endl;
+        }
     }
 
-    utils::g_log << "Hill climbing iterations: " << num_iterations << endl;
-    utils::g_log << "Hill climbing generated patterns: " << generated_patterns.size() << endl;
-    utils::g_log << "Hill climbing rejected patterns: " << num_rejected << endl;
-    utils::g_log << "Hill climbing maximum PDB size: " << max_pdb_size << endl;
-    utils::g_log << "Hill climbing time: "
-                 << hill_climbing_timer->get_elapsed_time() << endl;
+    if (log.is_at_least_normal()) {
+        log << "Hill climbing iterations: " << num_iterations << endl;
+        log << "Hill climbing generated patterns: " << generated_patterns.size() << endl;
+        log << "Hill climbing rejected patterns: " << num_rejected << endl;
+        log << "Hill climbing maximum PDB size: " << max_pdb_size << endl;
+        log << "Hill climbing time: "
+            << hill_climbing_timer->get_elapsed_time() << endl;
+    }
 
     delete hill_climbing_timer;
     hill_climbing_timer = nullptr;
 }
 
-PatternCollectionInformation PatternCollectionGeneratorHillclimbing::generate(
+string PatternCollectionGeneratorHillclimbing::name() const {
+    return "hill climbing pattern collection generator";
+}
+
+PatternCollectionInformation PatternCollectionGeneratorHillclimbing::compute_patterns(
     const shared_ptr<AbstractTask> &task) {
     TaskProxy task_proxy(*task);
     utils::Timer timer;
-    utils::g_log << "Generating patterns using the hill climbing generator..." << endl;
 
     // Generate initial collection: a pattern for each goal variable.
     PatternCollection initial_pattern_collection;
@@ -422,17 +448,16 @@ PatternCollectionInformation PatternCollectionGeneratorHillclimbing::generate(
     }
     current_pdbs = utils::make_unique_ptr<IncrementalCanonicalPDBs>(
         task_proxy, initial_pattern_collection);
-    utils::g_log << "Done calculating initial pattern collection: " << timer << endl;
+    if (log.is_at_least_normal()) {
+        log << "Done calculating initial pattern collection: " << timer << endl;
+    }
 
     State initial_state = task_proxy.get_initial_state();
     if (!current_pdbs->is_dead_end(initial_state) && max_time > 0) {
         hill_climbing(task_proxy);
     }
 
-    PatternCollectionInformation pci = current_pdbs->get_pattern_collection_information();
-    dump_pattern_collection_generation_statistics(
-        "Hill climbing generator", timer(), pci);
-    return pci;
+    return current_pdbs->get_pattern_collection_information();
 }
 
 void add_hillclimbing_options(OptionParser &parser) {
@@ -528,6 +553,7 @@ void add_hillclimbing_options(OptionParser &parser) {
         "infinity",
         Bounds("0.0", "infinity"));
     utils::add_rng_options(parser);
+    add_generator_options_to_parser(parser);
 }
 
 void check_hillclimbing_options(