+++ src/search/pdbs/pattern_collection_generator_manual.cc
@@ -14,12 +14,19 @@ using namespace std;
 
 namespace pdbs {
 PatternCollectionGeneratorManual::PatternCollectionGeneratorManual(const Options &opts)
-    : patterns(make_shared<PatternCollection>(opts.get_list<Pattern>("patterns"))) {
+    : PatternCollectionGenerator(opts),
+      patterns(make_shared<PatternCollection>(opts.get_list<Pattern>("patterns"))) {
 }
 
-PatternCollectionInformation PatternCollectionGeneratorManual::generate(
+string PatternCollectionGeneratorManual::name() const {
+    return "manual pattern collection generator";
+}
+
+PatternCollectionInformation PatternCollectionGeneratorManual::compute_patterns(
     const shared_ptr<AbstractTask> &task) {
-    utils::g_log << "Manual pattern collection: " << *patterns << endl;
+    if (log.is_at_least_normal()) {
+        log << "Manual pattern collection: " << *patterns << endl;
+    }
     TaskProxy task_proxy(*task);
     return PatternCollectionInformation(task_proxy, patterns);
 }
@@ -29,6 +36,7 @@ static shared_ptr<PatternCollectionGenerator> _parse(OptionParser &parser) {
         "patterns",
         "list of patterns (which are lists of variable numbers of the planning "
         "task).");
+    add_generator_options_to_parser(parser);
 
     Options opts = parser.parse();
     if (parser.dry_run())