+++ src/search/pdbs/pattern_collection_generator_multiple.h
@@ -8,13 +8,11 @@
 
 namespace options {
 class OptionParser;
-class Options;
 }
 
 namespace utils {
 class CountdownTimer;
 class RandomNumberGenerator;
-enum class Verbosity;
 }
 
 namespace pdbs {
@@ -36,14 +34,12 @@ namespace pdbs {
   blacklisting when stagnating.
 */
 class PatternCollectionGeneratorMultiple : public PatternCollectionGenerator {
-    const std::string name;
     const int max_pdb_size;
     const double pattern_generation_max_time;
     const double total_max_time;
     const double stagnation_limit;
     const double blacklisting_start_time;
     const bool enable_blacklist_on_stagnation;
-    const utils::Verbosity verbosity;
     std::shared_ptr<utils::RandomNumberGenerator> rng;
     const int random_seed;
 
@@ -63,7 +59,7 @@ class PatternCollectionGeneratorMultiple : public PatternCollectionGenerator {
     bool collection_size_limit_reached() const;
     bool time_limit_reached(const utils::CountdownTimer &timer) const;
     bool check_for_stagnation(const utils::CountdownTimer &timer);
-protected:
+    virtual std::string id() const = 0;
     virtual void initialize(const std::shared_ptr<AbstractTask> &task) = 0;
     virtual PatternInformation compute_pattern(
         int max_pdb_size,
@@ -72,11 +68,12 @@ protected:
         const std::shared_ptr<AbstractTask> &task,
         const FactPair &goal,
         std::unordered_set<int> &&blacklisted_variables) = 0;
-public:
-    PatternCollectionGeneratorMultiple(options::Options &opts, const std::string &name);
-    virtual ~PatternCollectionGeneratorMultiple() override = default;
-    virtual PatternCollectionInformation generate(
+    virtual std::string name() const override;
+    virtual PatternCollectionInformation compute_patterns(
         const std::shared_ptr<AbstractTask> &task) override;
+public:
+    explicit PatternCollectionGeneratorMultiple(options::Options &opts);
+    virtual ~PatternCollectionGeneratorMultiple() override = default;
 };
 
 extern void add_multiple_algorithm_implementation_notes_to_parser(