+++ src/search/pdbs/pattern_collection_generator_systematic.h
@@ -17,10 +17,6 @@ namespace causal_graph {
 class CausalGraph;
 }
 
-namespace options {
-class Options;
-}
-
 namespace pdbs {
 class CanonicalPDBsHeuristic;
 
@@ -44,11 +40,11 @@ class PatternCollectionGeneratorSystematic : public PatternCollectionGenerator {
     void build_sga_patterns(const TaskProxy &task_proxy, const causal_graph::CausalGraph &cg);
     void build_patterns(const TaskProxy &task_proxy);
     void build_patterns_naive(const TaskProxy &task_proxy);
+    virtual std::string name() const override;
+    virtual PatternCollectionInformation compute_patterns(
+        const std::shared_ptr<AbstractTask> &task) override;
 public:
     explicit PatternCollectionGeneratorSystematic(const options::Options &opts);
-
-    virtual PatternCollectionInformation generate(
-        const std::shared_ptr<AbstractTask> &task) override;
 };
 }