+++ src/search/pdbs/pattern_generator_random.cc
@@ -18,21 +18,19 @@ using namespace std;
 
 namespace pdbs {
 PatternGeneratorRandom::PatternGeneratorRandom(options::Options &opts)
-    : max_pdb_size(opts.get<int>("max_pdb_size")),
+    : PatternGenerator(opts),
+      max_pdb_size(opts.get<int>("max_pdb_size")),
       max_time(opts.get<double>("max_time")),
       bidirectional(opts.get<bool>("bidirectional")),
-      verbosity(opts.get<utils::Verbosity>("verbosity")),
       rng(utils::parse_rng_from_options(opts)) {
 }
 
-PatternInformation PatternGeneratorRandom::generate(
-    const shared_ptr<AbstractTask> &task) {
-    if (verbosity >= utils::Verbosity::NORMAL) {
-        utils::g_log << "Generating pattern using the Random Pattern algorithm."
-                     << endl;
-    }
+string PatternGeneratorRandom::name() const {
+    return "random pattern generator";
+}
 
-    utils::Timer timer;
+PatternInformation PatternGeneratorRandom::compute_pattern(
+    const shared_ptr<AbstractTask> &task) {
     vector<vector<int>> cg_neighbors = compute_cg_neighbors(
         task, bidirectional);
     TaskProxy task_proxy(*task);
@@ -41,20 +39,13 @@ PatternInformation PatternGeneratorRandom::generate(
     Pattern pattern = generate_random_pattern(
         max_pdb_size,
         max_time,
-        verbosity,
+        log,
         rng,
         task_proxy,
         goals[0].var,
         cg_neighbors);
 
-    PatternInformation result(task_proxy, pattern);
-    if (verbosity >= utils::Verbosity::NORMAL) {
-        dump_pattern_generation_statistics(
-            "Random Pattern",
-            timer.stop(),
-            result);
-    }
-    return result;
+    return PatternInformation(task_proxy, pattern);
 }
 
 static shared_ptr<PatternGenerator> _parse(options::OptionParser &parser) {
@@ -78,7 +69,7 @@ static shared_ptr<PatternGenerator> _parse(options::OptionParser &parser) {
         "infinity",
         Bounds("0.0", "infinity"));
     add_random_pattern_bidirectional_option_to_parser(parser);
-    utils::add_log_options_to_parser(parser);
+    add_generator_options_to_parser(parser);
     utils::add_rng_options(parser);
 
     Options opts = parser.parse();