+++ src/search/utils/logging.cc
@@ -15,8 +15,9 @@ namespace utils {
 /*
   NOTE: When adding more options to Log, make sure to adapt the if block in
   get_log_from_options below to test for *all* default values used for
-  global_log here.
+  global_log here. Also add the options to dump_options().
 */
+
 static shared_ptr<Log> global_log = make_shared<Log>(Verbosity::NORMAL);
 
 LogProxy g_log(global_log);